voice test/requirements_with_voice.txt
streamlit
pandas
scikit-learn
matplotlib
SpeechRecognition
streamlit-audiorecorder
PyAudio